Update MediaWiki 1.42.1 -> 1.42.6
Also:
- Migrate to buildComposerProject2
- Manually merge composer.json file; the plugin is spotty
- Drop finalAttrs in wiki for now, it's tricky and not used

I have slight paranoia about the composer.lock diff. It raises the question of why bacon/bacon-qr-code et al. were included in the vendor bundle, since none of the composer.json files we should be including appear to reference it. I was unable to trace what happened here. However, in testing, I don't see any ill effects as a result of this so far. Knock-on-wood.